fix: use BigInteger for telegram_id in User model

+++ backend/app/models/user.py
@@ -12,7 +12,7 @@ class UserRole(str, enum.Enum):
 class User(Base):
     __tablename__ = "users"
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, index=True)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, index=True)
     telegram_id: Mapped[int] = mapped_column(BigInteger, unique=True, index=True, nullable=False)
     
     username: Mapped[Optional[str]] = mapped_column(String(32), nullable=True)
